Split GitHub preview service modules
