+++ scripts/debt-ratchet-baseline.json
@@ -4,7 +4,7 @@
   "counts": {
     "any_casts": 4,
     "todo_comments": 0,
-    "oversized_files": 88,
+    "oversized_files": 87,
     "bootstrap_imports": 0,
     "redux_imports": 0,
     "pages_router_files": 0
@@ -12,7 +12,6 @@
   "oversized_file_allowlist": [
     "app/about/press/page.tsx",
     "app/api/farcaster/route.ts",
-    "app/api/github-preview/service.ts",
     "app/api/og-metadata/drops/[id]/image.tsx",
     "app/api/open-graph/6529/service.ts",
     "app/api/open-graph/compound/service.ts",
